pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "stubllm"
version = "0.1.0"
description = "Deterministic mock server for LLM APIs. Test your AI code without spending tokens."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [{ name = "mockllm contributors" }]
keywords = ["llm", "mock", "stub", "testing", "openai", "anthropic", "gemini"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
    "httpx>=0.26.0",
    "click>=8.1.0",
    "anyio>=4.0.0",
]

[project.optional-dependencies]
pytest = ["pytest>=7.0", "pytest-asyncio>=0.23.0"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0",
    "mypy>=1.8.0",
    "ruff>=0.3.0",
    "httpx>=0.26.0",
    "openai>=1.0.0",
]

[project.scripts]
stubllm = "stubllm.cli:cli"

[project.entry-points."pytest11"]
stubllm = "stubllm.pytest_plugin.plugin"

[tool.hatch.build.targets.wheel]
packages = ["src/stubllm"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "--cov --cov-report=term-missing --cov-fail-under=80"

[tool.mypy]
strict = true
python_version = "3.10"
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.coverage.run]
source_pkgs = ["stubllm"]
omit = ["*/tests/*"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
omit = [
    "*/__init__.py",
]
